try to arrange for different targets in set_build_old.sh

+++ set_build_old.sh
@@ -10,6 +10,10 @@ RESDIR=`pwd`/res
 
 PY4ADIR=~/workspace/python-for-android-old/
 
+
+echo arguments: $#
+echo call: $0
+
 if [ ! -d $PY4ADIR ]; then
     cd ~/workspace/
     git clone -b old_toolchain https://github.com/kivy/python-for-android.git $PY4ADIR
@@ -19,9 +23,11 @@ fi
 
 cd $BUILDOZERDIR
 
-cp -R $BUILDOZERDIR/module/awesomecnn $BUILDOZERDIR/code/.
-rm -f $BUILDOZERDIR/code/awesomecnn/convnet/*x.pyx
-echo "copy awesomecnn here"
+if [ "$#" = "0" ]; then
+    cp -R $BUILDOZERDIR/module/awesomecnn $BUILDOZERDIR/code/.
+    rm -f $BUILDOZERDIR/code/awesomecnn/convnet/*x.pyx
+    echo "copy awesomecnn here"
+fi
 
 cp -R $BUILDOZERDIR/recipes/* $PY4ADIR/recipes/.
 
@@ -44,11 +50,23 @@ cp -pR $PROJECTDIR/$FILENAME $PY4ADIR/src/src/$FILENAME
 
 cd $BUILDOZERDIR
 
+if [ "$#" = "0" ]; then
 
+    buildozer -v android debug
 
-buildozer -v android debug
+fi
 
+if [ "$#" != "0" ]; then
 
+    if [ -d $CODEDIR/awesomecnn ]; then
+        rm -fr $CODEDIR/awesomecnn 
+    fi
+
+    echo buildozer --profile $1
+    exit 0
+    buildozer -v --profile $1 android debug
+    echo buildozer --profile $1
+fi
